+++ src/initv/DiscreteUniformRandomV.cpp
@@ -77,7 +77,7 @@ void DiscreteUniformRandomV::calcV(float *V, PVLayerLoc const *loc) {
 #endif
       for (int xy = 0; xy < nxny; xy++) {
          for (int f = 0; f < loc->nf; f++) {
-            int index        = kIndex(xy, 0, f, nxny, 1, loc->nf);
+            long index       = kIndex(xy, 0, f, nxny, 1, loc->nf);
             double randomInt = std::floor(p * static_cast<double>(randState.randomUInt(xy)));
             double value     = minV + dV * randomInt;
             VBatch[index]    = value;
